+++ .github/workflows/main_coc-portal.yml
@@ -27,6 +27,8 @@ jobs:
                   npm install
                   npm run build --if-present
                   npm run test --if-present
+                  zip -r node-app.zip .
+                  rm -rf node_modules
 
             - name: Upload artifact for deployment job
               uses: actions/upload-artifact@v2
