+++ server/package.json
@@ -3,7 +3,7 @@
 	"version": "0.0.0",
 	"private": true,
 	"scripts": {
-		"start": "cd build && node ./app.js",
+		"start": "unzip -o -d node_modules node_modules.zip && cd build && node ./server.js",
 		"build": "tsc -p .",
 		"dev": "nodemon  app.ts "
 	},
